Phase 10: hygiene, archetype defects, a downwash map, and provenance
From the six role-play reviews (tasks/feedback/).

Hygiene: doc paths after the restructure, the vtol shim import, gusts
exported from cascade.env with public filter names, observation_size and
observation_layout, reset and step returning the same end of the delay
buffer, no float32 casts (dtype policy stated: JAX's default, x64 works),
diagnostic coefficients reported against a 1 m/s floor, trim channel
bounds from the mapped physical limits instead of a hard-coded unit, a
vertical mean-wind component, and the Dryden-at-hover validity note.

Archetypes: the parts mass is scaled to the aircraft mass (it summed to
1.06 to 1.08 of it) and the pod is a slender ellipsoid, not a sphere. The
static downwash fold, which scaled the tail's pitch damping and elevator
power along with the neutral point, is replaced by a real downwash map.

Physics: an optional S x S downwash_map in the spec and model. Surfaces
are evaluated in two passes, the upstream lift at the geometric incidence
and then every surface at its effective incidence, so a tail sees the
wing's actual lift and a stalled wing's collapsing lift takes the
downwash with it. A zero map reproduces the old single pass bitwise.

Model validation: the inertia inverse check allows float32 precision
(atol 1e-5; it rejected a physical 1 kg m^2 tensor) and principal
moments must obey the triangle inequality.

Provenance: cascade.provenance.stamp records package version, git commit,
JAX and jaxlib versions, backend, platform, x64, spec and model hashes,
and the seed; CITATION.cff; scripts/benchmark_env.py and
scripts/archetype_statistics.py emit the documented tables; uv build
produces 0.2.0 wheels.

+++ src/cascade/env/family.py
@@ -96,7 +96,7 @@ def sample_family(
         tasks=stack_pytrees(tasks),
         references=stack_pytrees(references),
         controllers=stack_pytrees(controllers),
-        cruise_speeds_m_s=jnp.asarray(np.asarray(speeds, dtype=np.float32)),
+        cruise_speeds_m_s=jnp.asarray(np.asarray(speeds, dtype=float)),
         designs=tuple(designs),
         reports=tuple(reports),
     )
